requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.0.10
django-crispy-forms==1.14.0
django-htmlmin==0.11.0
django-import-export==2.0.1
et-xmlfile==1.1.0
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==0.14.0
typing_extensions==4.0.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
